backend/pyproject.toml
[project]
name = "wonder-backend"
version = "0.1.0"
description = "Wonder audio DAW backend — merged canonical backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework & server
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "python-multipart>=0.0.9",

    # AI / ML
    "google-adk>=1.27.2",
    "google-generativeai>=0.8.0",

    # Vector database
    "lancedb>=0.22.0",

    # Audio processing
    "librosa>=0.10.0",
    "pydub>=0.25.1",
    "soundfile>=0.12",
    "numpy>=1.24",
    "scipy>=1.11",
    "pretty_midi>=0.2",
    "mido>=1.3",

    # Audio generation
    "elevenlabs>=1.0",

    # Database
    "motor>=3.3",
    "pymongo>=4.6",

    # Analytics
    "snowflake-connector-python>=3.0",

    # MCP
    "mcp[cli]>=1.6.0",

    # Config / utilities
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
stems = ["demucs>=4.0"]
midi = ["basic-pitch>=0.3"]
all = ["wonder-backend[stems,midi]"]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.4.0",
]

[tool.uv.workspace]
members = ["agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
